Extract podcast episodes to its own class
Removing it from the VC
